# test: optimize waitForNextUpstreamRequest()