Commit Message:
//test/integration:protocol_integration_test
After:
  Stats over 50 runs: max = 11.8s, min = 1.2s, avg = 4.0s, dev = 4.3s
Before:
  Stats over 50 runs: max = 2.6s, min = 1.4s, avg = 2.4s, dev = 0.3s

Signed-off-by: Yuchen Dai <silentdai@gmail.com>

Additional Description:
Risk Level: LOW
Testing: test
Docs Changes: n/a
Release Notes: n/a